Add types for each estimation method's parameters. Moreover, if the type is numeric (integer or float), a minimum and maximum value can optionally be set. Finally, each parameter has a default value.

+++ COACH/estimation_method/cocomo/COCOMO.py
@@ -21,7 +21,7 @@ class COCOMO(EstimationMethodService):
         if len(parameters_dict) != 2 or len(properties_dict) != 1:
             raise RuntimeError("Provided parameters does not match with the ontology")
         
-        return int(parameters_dict["a"]) * int(properties_dict["KLOC"]) ** int(parameters_dict["b"])
+        return float(parameters_dict["a"]) * float(properties_dict["KLOC"]) ** float(parameters_dict["b"])
     
 if __name__ == '__main__':
     COCOMO(sys.argv[1]).run()
